+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Use updated message adapter to store task metadata.
+
 ## [v1.0.3] - 2018-07-26
 
 - Fixed location for pulling execution name.